requirements.txt
aiodns==3.6.1
aiofiles==23.2.1
aiogram==3.4.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
APScheduler==3.10.4
async-timeout==4.0.3
asyncio==4.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dotenv==0.9.9
exceptiongroup==1.3.1
fastapi==0.123.5
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.20.0
Flask==3.1.2
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.19.4
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
magic-filter==1.0.12
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
maxminddb==3.0.0
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.7.0
narwhals==2.13.0
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.11.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==10.4.0
playwright==1.58.0
propcache==0.4.1
proxybroker==0.3.2
psycopg2-binary==2.9.11
PyAutoGUI==0.9.54
pycares==4.11.0
pycparser==2.23
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pyee==11.1.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pyppeteer==2.0.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==21.0.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
ruff==0.14.7
selenium==4.39.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==0.50.0
tomlkit==0.12.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
ttkbootstrap==1.20.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
webbot==0.34
websocket-client==1.9.0
websockets==10.4
Werkzeug==3.1.5
wsproto==1.3.2
yarl==1.22.0
yt-dlp==2025.11.12
zipp==3.23.0
